# applications/portal/api/openapi.yaml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
...
  /api/experiments/{id}/tag/:
    post:
      operationId: addTagsExperiment
      description: 'This viewset automatically provides `list`, `create`, `retrieve`,

        `update` and `destroy` actions.'
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this experiment.
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Tags'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tags'
          description: ''
      tags:
      - api
  /api/experiments/{id}/upload-files/:
    post:
      operationId: uploadFilesExperiment
      description: 'This viewset automatically provides `list`, `create`, `retrieve`,

        `update` and `destroy` actions.'
      parameters:
      - name: id
        in: path
        required: true
        description: A unique integer value identifying this experiment.
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ExperimentFileUpload'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentFileUpload'
          description: ''
      tags:
      - api
...
components:
  schemas:
...
    Tags:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string
      required:
      - tags
    ExperimentFileUpload:
      type: object
      properties:
        population_id:
          type: integer
        key_file:
          type: string
          format: binary
        data_file:
          type: string
          format: binary
      required:
      - key_file
      - data_file
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- bearerAuth: []